@@ -12,6 +12,6 @@ Gem::Specification.new do |s|
   s.platform     = Gem::Platform::RUBY
   s.files        = Dir.glob("{examples,lib,spec}/**/*") + %w[LICENSING README Rakefile README.rdoc]
 
-  s.add_dependency              'libxml-ruby', '~> 1.1.3'
-  s.add_development_dependency  'rspec',       '~> 1.3.0'
+  s.add_dependency              'libxml-ruby', '~> 2.8.0'
+  s.add_development_dependency  'rspec',       '~> 3.4.0'
 end
