@@ -1,7 +1,8 @@
 # Constants and descriptions for important elements/attributes
 # of the JMdict XML dictionary.                               
 # Descriptions come from JMdict.dtd (document type definition)
-module JMDictConstants
+module JDict
+  module JMDictConstants
     
     # TODO: change these strings to symbols ?
     # XML elements of the JMDict file
@@ -14,14 +15,14 @@ module JMDictConstants
 
       # This element will contain a word or short phrase in Japanese 
       # which is written using at least one kanji. The valid characters are
-    	# kanji, kana, related characters such as chouon and kurikaeshi, and
-    	# in exceptional cases, letters from other alphabets.
+      # kanji, kana, related characters such as chouon and kurikaeshi, and
+      # in exceptional cases, letters from other alphabets.
       KANJI          = 'keb'
 
       # This element content is restricted to kana and related
-    	# characters such as chouon and kurikaeshi. Kana usage will be
-    	# consistent between the keb and reb elements; e.g. if the keb
-    	# contains katakana, so too will the reb.
+      # characters such as chouon and kurikaeshi. Kana usage will be
+      # consistent between the keb and reb elements; e.g. if the keb
+      # contains katakana, so too will the reb.
       KANA           = 'reb'
 
       # The sense element will record the translational equivalent
@@ -31,13 +32,13 @@ module JMDictConstants
       SENSE          = 'sense'
 
       # Part-of-speech information about the entry/sense. Should use 
-    	# appropriate entity codes.
+      # appropriate entity codes.
       PART_OF_SPEECH = 'pos'
 
       # Within each sense will be one or more "glosses", i.e. 
-    	# target-language words or phrases which are equivalents to the 
-    	# Japanese word. This element would normally be present, however it 
-    	# may be omitted in entries which are purely for a cross-reference.
+      # target-language words or phrases which are equivalents to the 
+      # Japanese word. This element would normally be present, however it 
+      # may be omitted in entries which are purely for a cross-reference.
       GLOSS          = 'gloss'
     end
     
@@ -46,7 +47,7 @@ module JMDictConstants
       # Attribute specificying the gloss language
       LANGUAGE       = 'g_lang'
     end
-  
+    
     # Constants for selecting the search language.   
     # Used in the "gloss" element's g_lang attribute.
     #   :en never appears as a g_lang constant because gloss is assumed to be English when not specified
@@ -59,4 +60,5 @@ module JMDictConstants
       GERMAN   = :de
       RUSSIAN  = :ru
     end
+  end
 end
