@@ -1,2 +1,4 @@
-class Kana
+module JDict
+  class Kana
+  end
 end
