@@ -1,2 +1,4 @@
-class Kanji
+module JDict
+  class Kanji
+  end
 end
